+++ R/gg_pointrange.R
@@ -18,6 +18,7 @@
 #' @param pal_na Colour to use for NA values. A character vector of a hex code (or name).
 #' @param alpha Opacity. A number between 0 and 1.
 #' @param size Size. A number 0 upwards.
+#' @param fatten A multiplicative factor used to increase the size of the point.
 #' @param ... Other arguments passed to the relevant ggplot2::geom_* function.
 #' @param title Title string.
 #' @param subtitle Subtitle string.
@@ -92,6 +93,7 @@ gg_pointrange <- function(data = NULL,
                           pal_na = "#7F7F7F",
                           alpha = 1,
                           size = 0.5,
+                    fatten = 1,
                           ...,
                           title = NULL,
                           subtitle = NULL,
@@ -564,6 +566,7 @@ gg_pointrange <- function(data = NULL,
       position = position,
       alpha = alpha,
       size = size,
+      fatten = fatten,
       ...
     )