+++ NEWS.md
@@ -1,3 +1,10 @@
+# ggblanket 1.1.9000
+
+* Support for ggplotly.
+* New add_plotly_text function.
+* Updated theme title vjust and margins.
+* Fixed bug in gg_sf.
+
 # ggblanket 1.1.0
 
 * Remove default sentence case transformation of categorical variable labels.